Make it possible to toggle files w/o restarting. (#1)
Moved wavFile into MyHomePageState and added
a DropdownMenu into the AppBar as a (hacky)
way to toggle between the current two sound
files.

I have not yet fixed the loading of the sound
files to either use Flutter's asset system
or some other way which is more maintainable
than absolute paths.

Absolute paths are currently necessary since
when running the application it has no idea
where the sources are kept on disk and the
sound files are not currently bundled into
the application (asset system) nor are we
pulling them from a known shared location
(e.g. http to github, etc.)

I did not add tests in this PR, but maybe
we should do that on the show? :)